Add tests in ReservationControllerTest

+++ lib/common/src/main/java/nl/tudelft/sem11b/data/models/ReservationModel.java
@@ -3,12 +3,12 @@ package nl.tudelft.sem11b.data.models;
 import java.sql.Timestamp;
 
 public class ReservationModel {
-    private final int roomId;
+    private final long roomId;
     private final Timestamp since;
     private final Timestamp until;
     private final String title;
 
-    public int getRoomId() {
+    public long getRoomId() {
         return roomId;
     }
 
@@ -31,7 +31,7 @@ public class ReservationModel {
      * @param until end time of reservation
      * @param title title of reservation
      */
-    public ReservationModel(int roomId, Timestamp since, Timestamp until, String title) {
+    public ReservationModel(long roomId, Timestamp since, Timestamp until, String title) {
         this.roomId = roomId;
         this.since = since;
         this.until = until;
